Refactor income and expense modules and restructure routes and repositories
- Split income and expense logic into separate modules: sale, other, and buy
- Removed deprecated routes, services, schemas, and repositories
- Added new routes and handlers for income_sale, income_other, expense_buy, and expense_other
- Updated models, schemas, and repositories for pay and related entities
- Refactored dependency injection and tenant container setup
- Cleaned up unused movement_type components
- Improved consistency across services and route registration

+++ internal/schemas/product.go
@@ -63,7 +63,7 @@ type ProductCreate struct {
 	Code        string   `json:"code" validate:"required" example:"ABC123"`
 	Name        string   `json:"name" validate:"required" example:"Producto1"`
 	Description *string  `json:"description" example:"description|null"`
-	CategoryID  uint     `json:"category_id" validate:"required" example:"1"`
+	CategoryID  int64     `json:"category_id" validate:"required" example:"1"`
 	Price       *float64 `json:"price" example:"100.00"`
 	Notifier    bool     `json:"notifier" example:"false"`
 	MinAmount   float64  `json:"min_amount" example:"10.00"`
